added checks for authentication

+++ controllers/feed.js
@@ -1,6 +1,7 @@
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     
@@ -39,24 +40,29 @@ exports.createPost = (req, res, next) => {
     const title = req.body.title;
     const content = req.body.content;
     //const imageUrl = req.file.path;
+    let creator;
 
     const post = new Post({
         title: title, 
         content: content,
         imageUrl: 'images/image.jpg',
-        creator: {
-            name: 'Luke'
-        }
+        creator: req.userId
     });
     
     post.save()
+    .then(result => { 
+        return User.findById(req.userId)
+    })    
+    .then(user => {        
+        creator = user;
+        user.posts.push(post);
+        return user.save();        
+    })
     .then(result => {
-        
-        console.log(result)
-
         res.status(201).json({
             message: "Post created successfully!",
-            post: result
+            post: post, 
+            creator: { id: creator._id, name: creator.name }
         });
     })
     .catch(err => {
@@ -125,6 +131,12 @@ exports.updatePost = (req, res, next) => {
             throw error;
         }
 
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorized!');
+            error.statusCode = 403;
+            throw error;
+        }
+
         post.title = title;
         post.content = content;
         post.imageUrl = imageUrl;
@@ -162,7 +174,20 @@ exports.deletePost = (req, res, next) => {
             throw error;
         }
 
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error('Not authorized!');
+            error.statusCode = 403;
+            throw error;
+        }
+
        return Post.findByIdAndRemove(postId);
+    })
+    .then(result => {
+        return User.findById(req.userId);
+    })
+    .then(user => {
+    user.posts.pull(postId);
+    return user.save();
     })
     .then(result => {
